pyproject.toml
[project]
name = "pagemonk"
version = "0.1.0"
description = "AI-powered document processing and extraction platform"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "docling>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "ollama>=0.1.7",
    "python-magic>=0.4.27",
    "aiofiles>=23.2.0"
]
